requirements.txt
cffi
discord.py
pylinkirc
syncer
disco-py[performance,sharding]
requests[security]
aenum
mistletoe
